move exceptions to dedicated package
