+++ migrating-to-v3.md
@@ -4,17 +4,18 @@
 
  - Package was changed from `com.kttdevelopment.mal4j` to `dev.katsute.mal4j`.
  - Remove deprecated classes
-    - AndroidCompatibilityException
+    - `AndroidCompatibilityException`
  - Removed deprecated fields
-    - AnimeSource.Web_Manga (use AnimeSource.WebManga)
-    - AnimeSource.Digital_Manga (use AnimeSource.DigitalManga)
-    - MangaType.Novel (use MangaType.LightNovel)
+    - `AnimeSource.Web_Manga` (use `AnimeSource.WebManga`)
+    - `AnimeSource.Digital_Manga` (use `AnimeSource.DigitalManga`)
+    - `MangaType.Novel` (use `MangaType.LightNovel`)
  - Remove deprecated methods
-    - MyAnimeList.withOAuthToken (use MyAnimeList.withToken)
-    - MyAnimeList.withAuthorization (use MyAnimeList.withOAuth2)
-    - MyAnimeList.refreshOAuthToken (use MyAnimeList.refreshToken)
-    - MyAnimeList.getMyself (use MyAnimeList.getAuthenticatedUser)
-    - MyAnimeListAuthenticator String constructors (use Authorization constructors)
+    - `MyAnimeList.withOAuthToken` (use `MyAnimeList.withToken`)
+    - `MyAnimeList.withAuthorization` (use `MyAnimeList.withOAuth2`)
+    - `MyAnimeList.refreshOAuthToken` (use `MyAnimeList.refreshToken`)
+    - `MyAnimeList.getMyself` (use `MyAnimeList.getAuthenticatedUser`)
+    - `MyAnimeListAuthenticator` `String` constructors (use `Authorization` constructors)
+ - Moved exceptions from `dev.katsute.mal4j` to `dev.katsute.mal4j.exception`
 
 ## ⭐ New Features
