+++ src/main/java/dev/katsute/mal4j/MyAnimeListImpl.java
@@ -21,6 +21,8 @@ package dev.katsute.mal4j;
 import dev.katsute.mal4j.anime.*;
 import dev.katsute.mal4j.anime.property.AnimeRankingType;
 import dev.katsute.mal4j.anime.property.time.Season;
+import dev.katsute.mal4j.exception.HttpException;
+import dev.katsute.mal4j.exception.InvalidTokenException;
 import dev.katsute.mal4j.forum.*;
 import dev.katsute.mal4j.manga.*;
 import dev.katsute.mal4j.manga.property.MangaRankingType;
